requirements.txt
asgiref==3.8.1
boolean.py==4.0
CacheControl==0.14.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cyclonedx-python-lib==8.9.0
defusedxml==0.7.1
Django==5.1.5
django-environ==0.10.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.1
drf-spectacular
filelock==3.18.0
h11==0.14.0
idna==3.10
inflection==0.5.1
license-expression==30.4.1
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.0
packageurl-python==0.16.0
packaging==24.2
pip-api==0.0.34
pip-requirements-parser==32.0.1
platformdirs==4.3.7
psycopg2==2.9.10
psycopg2-binary==2.9.10
py-serializable==1.1.2
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.2.2
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
ruff==0.9.3
sortedcontainers==2.4.0
sqlparse==0.5.3
toml==0.10.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
